+++ R/addGeoArrowPolygonLayer.R
@@ -3,6 +3,8 @@
 #'
 #' @param map the [mapgl::maplibre()] or [mapgl::mapboxgl()] map to add the layer to.
 #' @param data a `sf`, `wk`, `geos` or `SpatVector` `(MULTI)POLYGON` object.
+#' Ignored if `source` is supplied.
+#' @param source the `id` of a source previously added via [addSource()].
 #' @param file a valid local file path to a `geoarrow` or `geoparquet` file to be
 #' added to the map. Ignored if `data` is supplied.
 #' @param url a URL to a remotely hosted `geoarrow` or `geoparquet` file to be
@@ -174,6 +176,7 @@
 addGeoArrowPolygonLayer = function(
     map
     , data
+    , source
     , file
     , url
     , layer_id = "polygon"
@@ -194,6 +197,7 @@ addGeoArrowPolygonLayer = function(
 .addGeoArrowPolygonLayer = function(
     map
     , data
+    , source
     , file
     , url
     , layer_id = "polygon"
@@ -232,28 +236,17 @@ addGeoArrowPolygonLayer = function(
     widget = map
   )
 
-  if (!missing(data)) {
-
-    data = parseGeoarrow(
-      data = data
-      , interleaved = TRUE
+  if (missing(source)) {
+    map = addSource(
+      map = map
+      , data = data
+      , file = file
+      , url = url
+      , id = layer_id
     )
-
+  } else {
+    layer_id = source
   }
-
-  map = geoarrowWidget::attachData(
-    widget = map
-    , data = data
-    , file = file
-    , url = url
-    , name = layer_id
-  )
-
-  extension_type = guessFileExtension(
-    data = data
-    , file = file
-    , url = url
-  )
 
   map$dependencies = c(
     map$dependencies
@@ -294,7 +287,6 @@ addGeoArrowPolygonLayer = function(
     , map_class = map_class
     , interleaved = TRUE
     , pickable = any(pickable(popup), pickable(tooltip))
-    , extension_type = extension_type
   )
 
   dot_lst = list(...)