+++ R/addGeoArrowScatterplotLayer.R
@@ -3,10 +3,12 @@
 #'
 #' @param map the [mapgl::maplibre()] or [mapgl::mapboxgl()] map to add the layer to.
 #' @param data a `sf`, `wk`, `geos` or `SpatVector` `(MULTI)POINT` object.
+#' Ignored if `source` is supplied.
+#' @param source the `id` of a source previously added via [addSource()].
 #' @param file a valid local file path to a `geoarrow` or `geoparquet` file to be
-#' added to the map. Ignored if `data` is supplied.
+#' added to the map. Ignored if `source` or `data` is supplied.
 #' @param url a URL to a remotely hosted `geoarrow` or `geoparquet` file to be
-#' added to the map. Ignored if `data` or `file` is supplied.
+#' added to the map. Ignored if `source` or `data` or `file` is supplied.
 #' @param layer_id the layer id.
 #' @param geom_column_name the name of the geometry column of data object.
 #' It is inferred automatically if only one is present.
@@ -237,6 +239,7 @@
 addGeoArrowScatterplotLayer = function(
     map
     , data
+    , source
     , file
     , url
     , layer_id = "scatter"
@@ -258,6 +261,7 @@ addGeoArrowScatterplotLayer = function(
 .addGeoArrowScatterplotLayer = function(
     map
     , data
+    , source
     , file
     , url
     , layer_id = "scatter"
@@ -272,8 +276,6 @@ addGeoArrowScatterplotLayer = function(
     , js_code
     , ...
 ) {
-
-  stopifnot(requireNamespace("geoarrow", quietly = TRUE))
 
   map$dependencies = c(
     map$dependencies
@@ -296,28 +298,17 @@ addGeoArrowScatterplotLayer = function(
     widget = map
   )
 
-  if (!missing(data)) {
-
-    data = parseGeoarrow(
-      data = data
-      , interleaved = TRUE
+  if (missing(source)) {
+    map = addSource(
+      map = map
+      , data = data
+      , file = file
+      , url = url
+      , id = layer_id
     )
-
+  } else {
+    layer_id = source
   }
-
-  map = geoarrowWidget::attachData(
-    widget = map
-    , data = data
-    , file = file
-    , url = url
-    , name = layer_id
-  )
-
-  extension_type = guessFileExtension(
-    data = data
-    , file = file
-    , url = url
-  )
 
   map$dependencies = c(
     map$dependencies
@@ -358,7 +349,6 @@ addGeoArrowScatterplotLayer = function(
     , map_class = map_class
     , interleaved = TRUE
     , pickable = any(pickable(popup), pickable(tooltip))
-    , extension_type = extension_type
   )
 
   dot_lst = list(...)