+++ inst/tinytest/test-addGeoArrowPathLayer.R
@@ -1,6 +1,6 @@
 # File created by roxut; edit the function definition file, not this file
 
-# Test found in addGeoArrowPathLayer.R:84 (file:line)
+# Test found in addGeoArrowPathLayer.R:86 (file:line)
   
 
 dummy_map <- structure(